Refactor useDebounce hook into own module
This starts us the hooks/ directory, where we'll collect our generic
custom hooks.
