+++ src/pages/search.js
@@ -9,6 +9,8 @@ import {
 } from 'react-instantsearch-dom';
 import qs from 'qs';
 
+import useDebounce from '../hooks/useDebounce';
+
 import Layout from '../components/layout';
 import SEO from '../components/seo';
 import DocumentHit from '../components/documentHit';
@@ -50,35 +52,6 @@ const urlToSearchState = ({ search }) => qs.parse(search.slice(1));
  * (milliseconds).
  */
 const DEBOUNCE_TIME = 800;
-
-/**
- * Debounce hook
- *
- * https://usehooks.com/useDebounce/
- */
-function useDebounce(value, delay) {
-  // State and setters for debounced value
-  const [debouncedValue, setDebouncedValue] = React.useState(value);
-
-  React.useEffect(
-    () => {
-      // Update debounced value after delay
-      const handler = setTimeout(() => {
-        setDebouncedValue(value);
-      }, delay);
-
-      // Cancel the timeout if value changes (also on delay change or unmount)
-      // This is how we prevent debounced value from updating if value is
-      // changed within the delay period. Timeout gets cleared and restarted.
-      return () => {
-        clearTimeout(handler);
-      };
-    },
-    [value, delay] // Only re-call effect if value or delay changes
-  );
-
-  return debouncedValue;
-}
 
 const AdvancedSearchPage = ({ location }) => {
   const [hitCardsExpanded, setHitCardsExpanded] = React.useState(false);
